+++ package.json
@@ -90,6 +90,7 @@
     "jest": "^29.0.1",
     "jest-axe": "^5.0.1",
     "jest-environment-jsdom": "^29.0.2",
+    "jest-fetch-mock": "^3.0.3",
     "node-fetch": "^2.6.7",
     "prettier": "^2.8.4",
     "prisma": "^4.3.1",
