+++ scripts/upsert/__tests__/courses.test.ts
@@ -1,7 +1,11 @@
+import { waitFor } from '@testing-library/react';
+
 import { differenceInDays, differenceInMinutes } from 'date-fns';
 
+import { Term } from 'lib/fetchers';
+
 import { createTask } from '../../../lib/task';
-import { upsertCoursesScript } from '../courses';
+import { upsertCoursesScript } from '../functions/upsertCourses';
 
 jest.mock('date-fns', () => ({
   differenceInMinutes: jest.fn(),
@@ -10,6 +14,7 @@ jest.mock('date-fns', () => ({
 
 jest.mock('../../../lib/task', () => ({
   findLatestTask: jest.fn(),
+  findLatestTaskByTerm: jest.fn((value: string, term: Term) => {}),
   createTask: jest.fn(),
 }));
 
@@ -23,13 +28,17 @@ describe('upsert-courses', () => {
   });
   it('should run the task every time if less than a week away from registration', async () => {
     (differenceInDays as jest.Mock).mockReturnValue(6);
-    await upsertCoursesScript('202009', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16));
+    await waitFor(() =>
+      upsertCoursesScript('202009', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16))
+    );
     expect(createTask).toBeCalledTimes(1);
   });
 
   it('should run the task every time if exactly a week away from registration', async () => {
     (differenceInDays as jest.Mock).mockReturnValue(7);
-    await upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16));
+    await waitFor(() =>
+      upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16))
+    );
     expect(createTask).toBeCalledTimes(1);
   });
 
@@ -37,28 +46,36 @@ describe('upsert-courses', () => {
     it('should run the task if before drop date and courses have not been updated within the last hour', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(61);
-      await upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(1);
     });
 
     it('should not run the task if before drop date and courses have been updated within the last hour', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(59);
-      await upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(0);
     });
 
     it('should run the task if after drop date and courses have not been updated within the last 24 hours', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(1441);
-      await upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(1);
     });
 
     it('should not run the task if after drop date and courses have been updated within the last 24 hours', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(1439);
-      await upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertCoursesScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(0);
     });
   });
