+++ scripts/upsert/__tests__/textbooks.test.ts
@@ -1,7 +1,9 @@
+import { waitFor } from '@testing-library/react';
+
 import { differenceInDays, differenceInMinutes } from 'date-fns';
 
 import { createTask } from '../../../lib/task';
-import { upsertTextbooksScript } from '../textbooks';
+import { upsertTextbooksScript } from '../functions/upsertTextbooks';
 
 jest.mock('date-fns', () => ({
   differenceInMinutes: jest.fn(),
@@ -10,6 +12,7 @@ jest.mock('date-fns', () => ({
 
 jest.mock('../../../lib/task', () => ({
   findLatestTask: jest.fn(),
+  findLatestTaskByTerm: jest.fn(),
   createTask: jest.fn(),
 }));
 
@@ -23,13 +26,17 @@ describe('upsert-textbooks', () => {
   });
   it('should run the task every time if less than a week away from registration', async () => {
     (differenceInDays as jest.Mock).mockReturnValue(6);
-    await upsertTextbooksScript('202009', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16));
+    await waitFor(() =>
+      upsertTextbooksScript('202009', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16))
+    );
     expect(createTask).toBeCalledTimes(1);
   });
 
   it('should run the task every time if exactly a week away from registration', async () => {
     (differenceInDays as jest.Mock).mockReturnValue(7);
-    await upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16));
+    await waitFor(() =>
+      upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 9, 16))
+    );
     expect(createTask).toBeCalledTimes(1);
   });
 
@@ -37,28 +44,36 @@ describe('upsert-textbooks', () => {
     it('should run the task if before drop date and courses have not been updated within the last hour', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(61);
-      await upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(1);
     });
 
     it('should not run the task if before drop date and courses have been updated within the last hour', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(59);
-      await upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 10, 31), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(0);
     });
 
     it('should run the task if after drop date and courses have not been updated within the last 24 hours', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(1441);
-      await upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(1);
     });
 
     it('should not run the task if after drop date and courses have been updated within the last 24 hours', async () => {
       (differenceInDays as jest.Mock).mockReturnValue(8);
       (differenceInMinutes as jest.Mock).mockReturnValue(1439);
-      await upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21));
+      await waitFor(() =>
+        upsertTextbooksScript('202101', new Date(2020, 9, 1), new Date(2020, 8, 20), new Date(2020, 8, 21))
+      );
       expect(createTask).toBeCalledTimes(0);
     });
   });
