+++ src/setupTests.ts
@@ -6,4 +6,9 @@ import '@testing-library/jest-dom';
 import '@testing-library/jest-dom/extend-expect';
 import { toHaveNoViolations } from 'jest-axe';
 
+import fetch from 'jest-fetch-mock';
+
 expect.extend(toHaveNoViolations);
+
+global.fetch = require('jest-fetch-mock');
+fetch.mockResponse(JSON.stringify({ testing: true }));
